requirements.txt
pandas==1.4.2
Pillow==9.4.0
plotly==5.6.0
streamlit==1.17.0
omero-marshal==0.8.0
sklearn3==0.0.1
scikit-learn==1.2.0